Added A Bridge Building Tool
To use the tool, either select the prefab in the folder or drag it into your scene.

On the object there's a custom inspector window for you to change bridge link/anchor prefabs (if you want different sprites) and a spot to type in the length of bridge you want. When done click build.